+++ src/main/java/wooteco/subway/dto/request/SectionRequest.java
@@ -1,4 +1,4 @@
-package wooteco.subway.dto;
+package wooteco.subway.dto.request;
 
 public class SectionRequest {
