+++ src/main/java/wooteco/subway/dto/response/PathResponse.java
@@ -1,4 +1,4 @@
-package wooteco.subway.dto;
+package wooteco.subway.dto.response;
 
 import java.util.ArrayList;
 import java.util.List;
