+++ src/main/java/wooteco/subway/service/LineService.java
@@ -8,10 +8,10 @@ import org.springframework.transaction.annotation.Isolation;
 import org.springframework.transaction.annotation.Transactional;
 import wooteco.subway.dao.LineDao;
 import wooteco.subway.domain.Line;
-import wooteco.subway.dto.LineRequest;
-import wooteco.subway.dto.LineResponse;
-import wooteco.subway.dto.SectionRequest;
-import wooteco.subway.dto.StationResponse;
+import wooteco.subway.dto.request.LineRequest;
+import wooteco.subway.dto.response.LineResponse;
+import wooteco.subway.dto.request.SectionRequest;
+import wooteco.subway.dto.response.StationResponse;
 import wooteco.subway.exception.DuplicateLineException;
 
 @Service
