+++ src/test/java/wooteco/subway/service/PathServiceTest.java
@@ -11,8 +11,8 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.context.jdbc.Sql;
 import org.springframework.transaction.annotation.Transactional;
-import wooteco.subway.dto.PathRequest;
-import wooteco.subway.dto.PathResponse;
+import wooteco.subway.dto.request.PathRequest;
+import wooteco.subway.dto.response.PathResponse;
 import wooteco.subway.exception.NotFoundStationException;
 
 @SpringBootTest
